+++ Team-3749-2025/src/main/java/frc/robot/subsystems/arm/climb/real/ClimbArmSparkMax.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.arm.climb;
+package frc.robot.subsystems.arm.climb.real;
 
 import com.revrobotics.spark.SparkAbsoluteEncoder;
 import com.revrobotics.spark.SparkBase.PersistMode;
@@ -9,6 +9,8 @@ import com.revrobotics.spark.config.SparkBaseConfig.IdleMode;
 import com.revrobotics.spark.config.SparkMaxConfig;
 import edu.wpi.first.math.MathUtil;
 import frc.robot.subsystems.arm.ArmConstants;
+import frc.robot.subsystems.arm.climb.ClimbArmIO;
+import frc.robot.subsystems.arm.climb.ClimbArmIO.ArmData;
 import frc.robot.utils.MiscConstants.SimConstants;
 
 /**
